pyproject.toml
[tool.poetry]
name = "zen3geo"
version = "0.2.0"
description = "The 🌏 data science library you've been waiting for~"
authors = ["Wei Ji <23487320+weiji14@users.noreply.github.com>"]
license = "LGPL-3.0-or-later"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.dependencies]
# Required
python = ">=3.8, <3.12"
rioxarray = ">=0.10.0"
torchdata = ">=0.4.0"
# Optional
datashader = {version = ">=0.14.0", optional = true}
pyogrio = {version = ">=0.4.0", extras = ["geopandas"], optional = true}
spatialpandas = {version = ">=0.4.0", optional = true}
xbatcher = {version = ">=0.1.0", optional = true}
# Docs
jupyter-book = {version="*", optional=true}
planetary-computer = {version="*", optional=true}
pystac = {version="*", optional=true}

[tool.poetry.group.dev.dependencies]
black = "*"
pytest = "*"

[tool.poetry.extras]
docs = [
    "datashader",
    "jupyter-book",
    "planetary-computer",
    "pyogrio",
    "pystac",
    "xbatcher"
]
raster = ["xbatcher"]
spatial = [
    "datashader",
    "spatialpandas"
]
vector = ["pyogrio"]

[tool.poetry-dynamic-versioning]
bump = true
enable = true
metadata = true
style = "pep440"

[build-system]
requires = ["poetry>=1.2.0b3", "poetry-dynamic-versioning"]
build-backend = "poetry.masonry.api"
